@@ -1,6 +1,6 @@
 class Spree::SuppliersController < Spree::StoreController
 
-  before_filter :check_if_supplier, only: [:create, :new]
+  before_action :check_if_supplier, only: [:create, :new]
   ssl_required
 
   def create
