@@ -15,15 +15,15 @@ module SpreeMarketplace
       end
 
       def add_migrations
-        run 'bundle exec rake railties:install:migrations FROM=spree_marketplace'
+        run 'bundle exec rails railties:install:migrations FROM=spree_marketplace'
       end
 
       def run_migrations
         run_migrations = options[:auto_run_migrations] || ['', 'y', 'Y'].include?(ask 'Would you like to run the migrations now? [Y/n]')
         if run_migrations
-          run 'bundle exec rake db:migrate'
+          run 'bundle exec rails db:migrate'
         else
-          puts 'Skipping rake db:migrate, don\'t forget to run it!'
+          puts 'Skipping rails db:migrate, don\'t forget to run it!'
         end
       end
     end
