@@ -1,4 +1,4 @@
 require 'stripe'
-require 'spree_core'
+require 'spree/core'
 require 'spree_drop_ship'
 require 'spree_marketplace/engine'
